requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.0
idna==3.4
lxml==4.9.3
multidict==6.0.4
nest-asyncio==1.5.8
packaging==23.2
pyshark==0.6
python-dotenv==1.0.0
termcolor==2.3.0
yarl==1.9.2
